review cleanup process and grant cleanup on fail
